Refactor AboutPage and OnboardingPage components to enhance structure and clarity. Introduced StickySectionToc for improved navigation, updated section titles, and removed outdated content. Streamlined the onboarding process by eliminating unnecessary state management and simplifying the layout.
